requirements.txt
PyYAML
h5py
tensorboardX
slot_attention
tqdm
wandb
matplotlib
sklearn